+++ libs/arcadeshield/_locales/arcadeshield-strings.json
@@ -24,7 +24,6 @@
   "{id:category}ScreenBitmap": "ScreenBitmap",
   "{id:category}Screenhelpers": "Screenhelpers",
   "{id:category}_helpers_workaround": "_helpers_workaround",
-  "{id:category}_screen_internal": "_screen_internal",
   "{id:group}Compare": "Compare",
   "{id:group}Create": "Create",
   "{id:group}Drawing": "Drawing",
