+++ libs/arcadeshield/pxt.json
@@ -23,31 +23,12 @@
         "bitmap.cpp",
         "bitmap.ts",
         "bitmap.d.ts",
-        "init.ts",
+        "bmp.ts",
         "fieldeditors.ts",
-        "ns.ts",
-        "frame.ts"
+        "ns.ts"
     ],
     "public": false,
     "preferredEditor": "tsprj",
-    "palette": [
-        "#000000",
-        "#ffffff",
-        "#ff2121",
-        "#ff93c4",
-        "#ff8135",
-        "#fff609",
-        "#249ca3",
-        "#78dc52",
-        "#003fad",
-        "#f2f2f2",
-        "#f2b233",
-        "#e45a33",
-        "#a5694f",
-        "#7c3f58",
-        "#91d2e7",
-        "#1e1e1e"
-    ],
     "disablesVariants": [
         "mbdal"
     ],
